refactor: generate credentials function
