+++ src/components/Publish/_utils.ts
@@ -13,7 +13,8 @@ import {
   getEventFromTx,
   ConsumerParameter,
   Metadata,
-  Service
+  Service,
+  Credentials
 } from '@oceanprotocol/lib'
 import { mapTimeoutStringToSeconds, normalizeFile } from '@utils/ddo'
 import { generateNftCreateData } from '@utils/nft'
@@ -96,6 +97,35 @@ export function transformConsumerParameters(
   return transformedValues as ConsumerParameter[]
 }
 
+export function generateCredentials(
+  oldCredentials: Credentials,
+  updatedAllow: string[],
+  updatedDeny: string[]
+): Credentials {
+  const updatedCredentials = {
+    allow: oldCredentials?.allow || [],
+    deny: oldCredentials?.deny || []
+  }
+
+  const credentialTypes = [
+    { type: 'allow', values: updatedAllow },
+    { type: 'deny', values: updatedDeny }
+  ]
+
+  credentialTypes.forEach((credentialType) => {
+    updatedCredentials[credentialType.type] = [
+      ...updatedCredentials[credentialType.type].filter(
+        (credential) => credential?.type !== 'address'
+      ),
+      ...(credentialType.values.length > 0
+        ? [{ type: 'address', values: credentialType.values }]
+        : [])
+    ]
+  })
+
+  return updatedCredentials
+}
+
 export async function transformPublishFormToDdo(
   values: FormPublishData,
   // Those 2 are only passed during actual publishing process
@@ -146,27 +176,6 @@ export async function transformPublishFormToDdo(
   const accessTermsFileInfo = gaiaXInformation.termsAndConditions
   const accessTermsUrlTransformed = accessTermsFileInfo?.length &&
     accessTermsFileInfo[0].valid && [sanitizeUrl(accessTermsFileInfo[0].url)]
-
-  const credentials = {
-    allow:
-      allow.length > 0
-        ? [
-            {
-              type: 'address',
-              values: allow
-            }
-          ]
-        : [],
-    deny:
-      deny.length > 0
-        ? [
-            {
-              type: 'address',
-              values: deny
-            }
-          ]
-        : []
-  }
 
   const newMetadata: Metadata = {
     created: currentTime,
@@ -249,6 +258,8 @@ export async function transformPublishFormToDdo(
       : undefined
   }
 
+  const newCredentials = generateCredentials(undefined, allow, deny)
+
   const newDdo: DDO = {
     '@context': ['https://w3id.org/did/v1'],
     id: did,
@@ -257,7 +268,7 @@ export async function transformPublishFormToDdo(
     chainId,
     metadata: newMetadata,
     services: [newService],
-    credentials,
+    credentials: newCredentials,
     // Only added for DDO preview, reflecting Asset response,
     // again, we can assume if `datatokenAddress` is not passed,
     // we are on preview.
